+++ plugins/sova-notifications/tests/notifications.rs
@@ -2,7 +2,8 @@
 
 use sova_core::{Json, Request, ResponseAssert, TestClient};
 use sova_notifications::{Channel, Notifications, NotificationsMigrator, Notify, Via};
-use sova_testing::{ActingAs, TestApp};
+use sova_notifications::testing::ActingAs;
+use sova_testing::TestApp;
 use serde_json::json;
 
 #[tokio::test]